Usage of loop and arguments object
You can eliminate the variable limit by looping over the arguments object, therefore offering a wider usage of variables in a string. :)

textToFormat will be at index 0 of the arguments, so you can start the iteration at 1, and subtract by one within the regexp so you can use {0} within the string literal. You may want to add a check in the loop to filter out variables that are not a string.